tower_governor::governor

Struct GovernorConfigBuilder

pub struct GovernorConfigBuilder<K: KeyExtractor, M: RateLimitingMiddleware<QuantaInstant>> { /* private fields */ }

Helper struct for building a configuration for the governor middleware.

Example

Create a configuration with a quota of ten requests per IP address that replenishes one element every minute.

use tower_governor::governor::GovernorConfigBuilder;

let config = GovernorConfigBuilder::default()
    .per_second(60)
    .burst_size(10)
    .finish()
    .unwrap();

with x-ratelimit headers

use tower_governor::governor::GovernorConfigBuilder;

let config = GovernorConfigBuilder::default()
    .per_second(60)
    .burst_size(10)
    .use_headers() // Add this
    .finish()
    .unwrap();

Implementations

impl<K: KeyExtractor, M: RateLimitingMiddleware<QuantaInstant>> GovernorConfigBuilder<K, M>

pub fn error_handler<F>(&mut self, func: F) -> &mut Self
where F: Fn(GovernorError) -> Response<Body> + Send + Sync + 'static,

Set handler function for handling GovernorError

Example

GovernorConfigBuilder::default()
    .error_handler(|mut error| {
        // match against GovernorError and produce customized Response type.
        match error {
            _ => Response::new("some error".into())
        }
    });

impl<M: RateLimitingMiddleware<QuantaInstant>> GovernorConfigBuilder<PeerIpKeyExtractor, M>

Sets the default Governor Config and defines all the different configuration functions This one is used when the default PeerIpKeyExtractor is used

pub fn const_default() -> Self

pub fn const_period(self, duration: Duration) -> Self

Set the interval after which one element of the quota is replenished.

The interval must not be zero.

pub fn const_per_second(self, seconds: u64) -> Self

Set the interval after which one element of the quota is replenished in seconds.

The interval must not be zero.

pub fn const_per_millisecond(self, milliseconds: u64) -> Self

Set the interval after which one element of the quota is replenished in milliseconds.

The interval must not be zero.

pub fn const_per_nanosecond(self, nanoseconds: u64) -> Self

Set the interval after which one element of the quota is replenished in nanoseconds.

The interval must not be zero.

pub fn const_burst_size(self, burst_size: u32) -> Self

Set quota size that defines how many requests can occur before the governor middleware starts blocking requests from an IP address and clients have to wait until the elements of the quota are replenished.

The burst_size must not be zero.

impl<K: KeyExtractor, M: RateLimitingMiddleware<QuantaInstant>> GovernorConfigBuilder<K, M>

Sets configuration options when any Key Extractor is provided

pub fn period(&mut self, duration: Duration) -> &mut Self

Set the interval after which one element of the quota is replenished.

The interval must not be zero.

pub fn per_second(&mut self, seconds: u64) -> &mut Self

Set the interval after which one element of the quota is replenished in seconds.

The interval must not be zero.

pub fn per_millisecond(&mut self, milliseconds: u64) -> &mut Self

Set the interval after which one element of the quota is replenished in milliseconds.

The interval must not be zero.

pub fn per_nanosecond(&mut self, nanoseconds: u64) -> &mut Self

Set the interval after which one element of the quota is replenished in nanoseconds.

The interval must not be zero.

pub fn burst_size(&mut self, burst_size: u32) -> &mut Self

Set quota size that defines how many requests can occur before the governor middleware starts blocking requests from an IP address and clients have to wait until the elements of the quota are replenished.

The burst_size must not be zero.

pub fn methods(&mut self, methods: Vec<Method>) -> &mut Self

Set the HTTP methods this configuration should apply to. By default this is all methods.

pub fn key_extractor<K2: KeyExtractor>( &mut self, key_extractor: K2, ) -> GovernorConfigBuilder<K2, M>

Set the key extractor this configuration should use. By default this is using the PeerIpKeyExtractor.

pub fn use_headers( &mut self, ) -> GovernorConfigBuilder<K, StateInformationMiddleware>

Set x-ratelimit headers to response, the headers is

x-ratelimit-limit - Request limit

x-ratelimit-remaining - The number of requests left for the time window

x-ratelimit-after - Number of seconds in which the API will become available after its rate limit has been exceeded

x-ratelimit-whitelisted - If the request method not in methods, this header will be add it, use methods to add methods

By default x-ratelimit-after is enabled, with use_headers will enable x-ratelimit-limit, x-ratelimit-whitelisted and x-ratelimit-remaining

pub fn finish(&mut self) -> Option<GovernorConfig<K, M>>

Finish building the configuration and return the configuration for the middleware. Returns None if either burst size or period interval are zero.