tower_governor

Module key_extractor

Structs

GlobalKeyExtractor

A KeyExtractor that allow to do rate limiting for all incoming requests. This is useful if you want to hard-limit the HTTP load your app can handle.

PeerIpKeyExtractor

A KeyExtractor that uses peer IP as key. This is the default key extractor and it may no do want you want.

SmartIpKeyExtractor

A KeyExtractor that tries to get the client IP address from the x-forwarded-for, x-real-ip, and forwarded headers in that order. Falls back to the peer IP address.

Traits

KeyExtractor

Generic structure of what is needed to extract a rate-limiting key from an incoming request.